feat: refactor tabs & add useTabs hook
